services/calculation-service/requirements.txt
Flask
laspy
pvlib
psycopg2-binary
python-dotenv
Shapely
